Split decompression out from frame and header handling.
